Initial implementation of Flat Citrinet (flattened Parallel Citrinet).
The implementation is not the best as I had to work around neural module typing system
to aggregate existing components together. As an example, that means that ParallelBlock can't be
directly used in implementation as it passes inputs without parameter names.

A better implementation should introduce intermediate non-typed class with Jasper block sequence
and use it to construct various encoders.

The config has not been modified from PCitrinet, so probably the results won't be great.

Signed-off-by: Aleksei Kalinov <alekseia@nvidia.com>